+++ src/main/kotlin/com/github/iml885203/intellijgitopen/actions/OpenRemoteRepoAction.kt
@@ -12,7 +12,7 @@ class OpenRemoteRepoAction : AnAction() {
         val project = e.project ?: return
         val projectPath = project.basePath ?: return
         if (!GitHelper.isGitRepository(projectPath)) {
-            MyNotifier.notifyError(project, "The current project is not a Git repository.")
+            MyNotifier.notifyWarn(project, "The current project is not a Git repository.")
             return
         }