+++ CHANGELOG.md
@@ -3,4 +3,21 @@
 # IntelliJ-git-open Changelog
 
 ## [Unreleased]
-### Added
+
+## [2024.6.5]
+- Fix Create Merge Request popup will switch to first project
+- Fix Create Merge Request popup sorting
+- Fix Create Merge Request popup filter current branch
+
+## [2024.6.4]
+- Support Create Merge Request
+
+## [2024.6.3]
+- Can open ssh remote url
+- Support Open Pipelines/Actions On Browser
+
+## [2024.6.2]
+- Support open merge request
+
+## [2024.6.1]
+- Support open remote repo